Refactored Bogey Style System
