+++ src/main/java/com/rabbitminers/extendedbogeys/bogey/styles/content/DefaultStyle.java
@@ -1,5 +1,6 @@
-package com.rabbitminers.extendedbogeys.bogey.styles;
+package com.rabbitminers.extendedbogeys.bogey.styles.content;
 
+import com.rabbitminers.extendedbogeys.bogey.styles.IBogeyStyle;
 import com.rabbitminers.extendedbogeys.bogey.util.LanguageKey;
 import com.simibubi.create.content.logistics.trains.track.TrackBlock;
 import com.simibubi.create.foundation.utility.Lang;
